First steps on creating an interface for handling waterfiltration. Tagged Airfiltration along for the ride.
Also fleshed out the MaxStackableQuantity a bit more so the CoffeeFilter can make use of it.

+++ GanjaLibrary/Classes/Items/Item.cs
@@ -12,7 +12,7 @@ namespace GanjaLibrary.Classes.Items
         public double Weight { get; set; }
         public double Value { get; set; }
 
-        public double MaxStackableQuantity { get; set; }
+        public int MaxStackableQuantity { get; set; }
 
         public Guid Id { get; set; }
